+++ rpc/include/rpc/stub.h
@@ -69,7 +69,7 @@ namespace rpc
     class i_interface_stub
     {
     public:
-        virtual interface_ordinal get_interface_id(uint8_t rpc_version) const = 0;
+        virtual interface_ordinal get_interface_id(uint64_t rpc_version) const = 0;
         virtual int call(uint64_t protocol_version, caller_channel_zone caller_channel_zone_id, caller_zone caller_zone_id, method method_id, size_t in_size_, const char* in_buf_, std::vector<char>& out_buf_)
             = 0;
         virtual int cast(interface_ordinal interface_id, shared_ptr<i_interface_stub>& new_stub) = 0;
