+++ rpc/include/rpc/version.h
@@ -6,11 +6,11 @@ namespace rpc
     //128 versions should be enough, if we need more we can use the eighth bit to carry to the next bit
 #ifndef NO_RPC_V1
     #define RPC_V1
-    constexpr std::uint8_t VERSION_1 = 1;
+    constexpr std::uint64_t VERSION_1 = 1;
 #endif
 #ifndef NO_RPC_V2
     #define RPC_V2
-    constexpr std::uint8_t VERSION_2 = 2;
+    constexpr std::uint64_t VERSION_2 = 2;
 #endif
-    std::uint8_t get_version();
+    std::uint64_t get_version();
 }
